Backend/requirements.txt
fastapi==0.100.1
fastapi_cache2[redis]==0.2.1
pydantic[email]==2.1.1
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
starlette==0.27.0
uvicorn==0.23.2
aiosqlite==0.19.0
websockets==11.0.3
alembic==1.11.1